+++ src/packages/map/api/MapModel.ts
@@ -39,6 +39,7 @@ export interface HighlightStyle {
     Point?: StyleLike;
     LineString?: StyleLike;
     Polygon?: StyleLike;
+    MultiPolygon?: StyleLike;
 }
 
 /**