+++ src/samples/map-sample/ol-app/build.config.mjs
@@ -7,17 +7,26 @@ export default defineBuildConfig({
     i18n: ["en", "de"],
     services: {
         MapConfigProviderImpl: {
-            provides: ["map.MapConfigProvider"]
+            provides: ["map.MapConfigProvider"],
+            references: {
+                vectorSourceFactory: "ogc-features.VectorSourceFactory"
+            }
         },
         AppConfig: {
             provides: "ol-app.AppConfig",
             references: {
                 ogcSearchSourceFactory: "ogc-features.SearchSourceFactory",
-                httpService: "http.HttpService"
+                httpService: "http.HttpService",
+                mapRegistry: "map.MapRegistry"
             }
         }
     },
     ui: {
-        references: ["map-sample-logging.LogService", "map.MapRegistry", "ol-app.AppConfig"]
+        references: [
+            "map-sample-logging.LogService",
+            "map.MapRegistry",
+            "ol-app.AppConfig",
+            "notifier.NotificationService"
+        ]
     }
 });