+++ src/packages/spatial-bookmarks/CHANGELOG.md
@@ -28,3 +28,32 @@
 -   Updated dependencies [0c4ce04]
     -   @open-pioneer/map@0.1.1
     -   @open-pioneer/local-storage@0.1.0
+
+## 0.2.0
+
+### Minor Changes
+
+-   70349a8: Update to new core packages major versions
+
+### Patch Changes
+
+-   Updated dependencies [70349a8]
+    -   @open-pioneer/local-storage@0.2.0
+    -   @open-pioneer/map@0.2.0
+    -   @open-pioneer/react-utils@0.2.0
+
+## 0.1.0
+
+### Minor Changes
+
+-   3ca8a25: Initial release
+
+### Patch Changes
+
+-   Updated dependencies [08bffbc]
+-   Updated dependencies [a58546b]
+-   Updated dependencies [3ca8a25]
+-   Updated dependencies [a58546b]
+-   Updated dependencies [0c4ce04]
+    -   @open-pioneer/map@0.1.1
+    -   @open-pioneer/local-storage@0.1.0